requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.4
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
gensim==4.3.2
greenlet==3.0.3
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.28.0
ipython==8.12.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.6.0
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langcodes==3.3.0
langsmith==0.0.75
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathy==0.10.3
pickleshare==0.7.5
Pillow==10.1.0
pip==23.3.2
platformdirs==4.1.0
preshed==3.0.9
prompt-toolkit==3.0.42
psutil==5.9.7
pure-eval==0.2.2
pydantic==1.10.13
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.2.1
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.2.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.24
srsly==2.4.8
stack-data==0.6.2
sympy==1.12
tenacity==8.2.3
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.2
typer==0.7.0
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.1.0
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
wheel==0.42.0
yarl==1.9.4
zipp==3.17.0